+++ sapling_proto/src/storage/sapling.rs
@@ -17,26 +17,26 @@ pub trait SaplingStorage {
     fn get_head(&mut self) -> Result<SaplingHead>;
 
     // Dual representation: as a ring queue and as a hashset
-    fn set_root(&mut self, root: Hash, position: usize) -> Result<()>;
+    fn set_root(&mut self, root: Hash, position: u64) -> Result<()>;
     fn has_root(&self, root: &Hash) -> Result<bool>;
-    fn get_root(&mut self, position: usize) -> Result<Option<Hash>>;
+    fn get_root(&mut self, position: u64) -> Result<Option<Hash>>;
 
     // Dual representation: as an array and as a hashset
-    fn set_nullifier(&mut self, nullifier: Nullifier, position: usize) -> Result<()>;
+    fn set_nullifier(&mut self, nullifier: Nullifier, position: u64) -> Result<()>;
     fn has_nullifier(&self, nullifier: &Nullifier) -> Result<bool>;
-    fn get_nullifier(&mut self, position: usize) -> Result<Option<Nullifier>>;
+    fn get_nullifier(&mut self, position: u64) -> Result<Option<Nullifier>>;
 
     // Flattened incremental Merkle tree
     // The height of the leaves level is 0, height of the root is [MAX_HEIGHT]
     // Path is the sequential number of a node in the tree: root = 1, left = 2, right = 3, etc
     // [CommitmentNode] is a more generic type for commitment tree structure, can be constructed out of a [Commitment]
-    fn set_commitment(&mut self, commitment: CommitmentNode, path: usize) -> Result<()>;
-    fn get_commitment(&mut self, path: usize) -> Result<Option<CommitmentNode>>;
+    fn set_commitment(&mut self, commitment: CommitmentNode, path: u64) -> Result<()>;
+    fn get_commitment(&mut self, path: u64) -> Result<Option<CommitmentNode>>;
 
     // Linked to leaves-level commitments
     // Position is relative number of the leaf, starts from 0, actually it's [Path] - 2 ^ [MAX_HEIGHT]
-    fn set_ciphertext(&mut self, ciphertext: Ciphertext, position: usize) -> Result<()>;
-    fn get_ciphertext(&mut self, position: usize) -> Result<Option<Ciphertext>>;
+    fn set_ciphertext(&mut self, ciphertext: Ciphertext, position: u64) -> Result<()>;
+    fn get_ciphertext(&mut self, position: u64) -> Result<Option<Ciphertext>>;
 
     fn commit(&mut self) -> Result<()>;
     fn rollback(&mut self);
@@ -53,7 +53,7 @@ impl<Backend: StoreBackend> SaplingStorage for LayeredStore<Backend> {
         Ok(self.set("/head".into(), Some(head))?)
     }
 
-    fn set_root(&mut self, root: Hash, position: usize) -> Result<()> {
+    fn set_root(&mut self, root: Hash, position: u64) -> Result<()> {
         if let Some(expired_root) = self.get_root(position)? {
             self.set::<Hash>(format!("/roots/index/{}", hex::encode(expired_root)), None)?;
         }
@@ -66,11 +66,11 @@ impl<Backend: StoreBackend> SaplingStorage for LayeredStore<Backend> {
         Ok(self.has(format!("/roots/index/{}", hex::encode(root)))?)
     }
 
-    fn get_root(&mut self, position: usize) -> Result<Option<Hash>> {
+    fn get_root(&mut self, position: u64) -> Result<Option<Hash>> {
         Ok(self.get(format!("/roots/list/{}", position))?)
     }
 
-    fn set_nullifier(&mut self, nullifier: Nullifier, position: usize) -> Result<()> {
+    fn set_nullifier(&mut self, nullifier: Nullifier, position: u64) -> Result<()> {
         self.set(format!("/nullifiers/list/{}", position), Some(nullifier.0))?;
         self.set(
             format!("/nullifiers/index/{}", hex::encode(nullifier.0)),
@@ -83,19 +83,19 @@ impl<Backend: StoreBackend> SaplingStorage for LayeredStore<Backend> {
         Ok(self.has(format!("/nullifiers/index/{}", hex::encode(nullifier.0)))?)
     }
 
-    fn get_nullifier(&mut self, position: usize) -> Result<Option<Nullifier>> {
+    fn get_nullifier(&mut self, position: u64) -> Result<Option<Nullifier>> {
         Ok(self
             .get(format!("/nullifiers/list/{}", position))?
             .map(|nf| Nullifier(nf)))
     }
 
-    fn set_commitment(&mut self, commitment: CommitmentNode, path: usize) -> Result<()> {
+    fn set_commitment(&mut self, commitment: CommitmentNode, path: u64) -> Result<()> {
         let mut cm = [0u8; 32];
         commitment.write(cm.as_mut_slice())?;
         Ok(self.set(format!("/commitments/{}", path), Some(cm))?)
     }
 
-    fn get_commitment(&mut self, path: usize) -> Result<Option<CommitmentNode>> {
+    fn get_commitment(&mut self, path: u64) -> Result<Option<CommitmentNode>> {
         if let Some(cm) = self.get::<[u8; 32]>(format!("/commitments/{}", path))? {
             Ok(Some(CommitmentNode::read(cm.as_slice())?))
         } else {
@@ -103,11 +103,11 @@ impl<Backend: StoreBackend> SaplingStorage for LayeredStore<Backend> {
         }
     }
 
-    fn set_ciphertext(&mut self, ciphertext: Ciphertext, position: usize) -> Result<()> {
+    fn set_ciphertext(&mut self, ciphertext: Ciphertext, position: u64) -> Result<()> {
         Ok(self.set(format!("/ciphertexts/{}", position), Some(ciphertext))?)
     }
 
-    fn get_ciphertext(&mut self, position: usize) -> Result<Option<Ciphertext>> {
+    fn get_ciphertext(&mut self, position: u64) -> Result<Option<Ciphertext>> {
         Ok(self.get(format!("/ciphertexts/{}", position))?)
     }
